Add files via upload
logic is pretty simple if 'x' container stores 'a' type of balls, then 'x' must trade all the 'a' type of balls from other containers at the cost of all non 'a' type balls it has. hence sum of non 'a' balls in 'x'=sum of all 'a' balls in other container adding no. of 'a' type in 'x' to both side. this becomes sum of xth row= ath column so all we have to do is to compare if the arrays containing all horizontal sums and vertical sums are same or not.